requirements.txt
behave
python-dotenv
allure-behave
selenium
webdriver-manager
selenium-wire
